+++ src/QuickMarcEditor/QuickMarcEditorRows/BytesField/BytesField.test.js
@@ -7,48 +7,38 @@ import { IntlProvider } from 'react-intl';
 
 import '@folio/stripes-acq-components/test/jest/__mock__';
 
-import FixedField, { SUBFIELD_TYPES } from './FixedField';
+import {
+  BytesField,
+  SUBFIELD_TYPES,
+} from './BytesField';
 
 const config = {
-  colSizes: [4, 4, 4],
   fields: [
-    [
-      undefined,
-      {
-        type: SUBFIELD_TYPES.BYTE,
-        name: 'Type',
-      },
-    ],
-    [
-      {
-        type: SUBFIELD_TYPES.STRING,
-        name: 'ELvl',
-      },
-      undefined,
-      {
-        type: SUBFIELD_TYPES.BYTE,
-        name: 'Relf',
-      },
-    ],
-    [
-      {
-        type: SUBFIELD_TYPES.BYTE,
-        name: 'Ctrl',
-      },
-      {
-        type: SUBFIELD_TYPES.BYTES,
-        name: 'Lang',
-        bytes: 3,
-      },
-    ],
+    {
+      type: SUBFIELD_TYPES.BYTE,
+      name: 'Type',
+    },
+    {
+      type: SUBFIELD_TYPES.STRING,
+      name: 'ELvl',
+    },
+    {
+      type: SUBFIELD_TYPES.BYTE,
+      name: 'Relf',
+    },
+    {
+      type: SUBFIELD_TYPES.BYTE,
+      name: 'Ctrl',
+    },
+    {
+      type: SUBFIELD_TYPES.BYTES,
+      name: 'Lang',
+      bytes: 3,
+    },
   ],
 };
-const fields = config.fields
-  .reduce((acc, row) => ([
-    ...acc,
-    ...row.filter(field => field),
-  ]), []);
-const messages = fields
+
+const messages = config.fields
   .reduce((acc, { name }) => ({ ...acc, [`ui-quick-marc.record.fixedField.${name}`]: name }), {});
 
 const renderFixedField = () => (render(
@@ -58,7 +48,7 @@ const renderFixedField = () => (render(
         onSubmit={jest.fn()}
         mutators={{ ...arrayMutators }}
         render={() => (
-          <FixedField
+          <BytesField
             name="records"
             config={config}
           />
@@ -68,27 +58,19 @@ const renderFixedField = () => (render(
   </IntlProvider>,
 ));
 
-describe('Given Fixed Field', () => {
+describe('Given Bytes Field', () => {
   afterEach(cleanup);
-
-  it('Than it should display all rows from passed config', () => {
-    const { getAllByTestId } = renderFixedField();
-
-    expect(getAllByTestId('fixed-field-row').length).toBe(config.fields.length);
-  });
 
   it('Than it should display all cols from passed config', () => {
     const { getAllByTestId } = renderFixedField();
 
-    const colLength = config.fields.reduce((acc, row) => acc + row.length, 0);
-
-    expect(getAllByTestId('fixed-field-col').length).toBe(colLength);
+    expect(getAllByTestId('bytes-field-col').length).toBe(config.fields.length);
   });
 
   it('Than it should display all fields from passed config', () => {
     const { getByText } = renderFixedField();
 
-    fields
+    config.fields
       .forEach(({ name }) => {
         expect(getByText(name)).toBeDefined();
       });
@@ -97,7 +79,7 @@ describe('Given Fixed Field', () => {
   it('Than it should display byte fields from passed config', () => {
     const { getAllByTestId } = renderFixedField();
 
-    const byteFields = fields.filter(({ type }) => type === SUBFIELD_TYPES.BYTE);
+    const byteFields = config.fields.filter(({ type }) => type === SUBFIELD_TYPES.BYTE);
 
     expect(getAllByTestId(`fixed-field-${SUBFIELD_TYPES.BYTE}`).length).toBe(byteFields.length);
   });
@@ -105,7 +87,7 @@ describe('Given Fixed Field', () => {
   it('Than it should display bytes fields from passed config', () => {
     const { getAllByTestId } = renderFixedField();
 
-    const bytesFieldsLength = fields
+    const bytesFieldsLength = config.fields
       .filter(({ type }) => type === SUBFIELD_TYPES.BYTES)
       .reduce((acc, field) => acc + field.bytes, 0);
 
@@ -115,7 +97,7 @@ describe('Given Fixed Field', () => {
   it('Than it should display string fields from passed config', () => {
     const { getAllByTestId } = renderFixedField();
 
-    const stringFields = fields.filter(({ type }) => type === SUBFIELD_TYPES.STRING);
+    const stringFields = config.fields.filter(({ type }) => type === SUBFIELD_TYPES.STRING);
 
     expect(getAllByTestId(`fixed-field-${SUBFIELD_TYPES.STRING}`).length).toBe(stringFields.length);
   });