+++ src/QuickMarcEditor/QuickMarcEditorRows/FixedField/index.js
@@ -1,5 +1,3 @@
-export { default as FixedField } from './FixedField';
-export * from './FixedField';
 export * from './FixedFieldFactory';
 
 export { default as BookFixedField } from './BookFixedField';