+++ src/QuickMarcEditor/QuickMarcEditorRows/MaterialCharsField/VisualMaterialCharsField.js
@@ -1,0 +1,54 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+import {
+  BytesField,
+  SUBFIELD_TYPES,
+} from '../BytesField';
+
+const config = {
+  fields: [
+    {
+      name: 'Ills',
+      type: SUBFIELD_TYPES.BYTES,
+      bytes: 4,
+    },
+    {
+      name: 'Audn',
+      type: SUBFIELD_TYPES.BYTE,
+    },
+    {
+      name: 'GPub',
+      type: SUBFIELD_TYPES.BYTE,
+    },
+    {
+      name: 'Form',
+      type: SUBFIELD_TYPES.BYTE,
+    },
+    {
+      name: 'TMat',
+      type: SUBFIELD_TYPES.BYTE,
+    },
+    {
+      name: 'Tech',
+      type: SUBFIELD_TYPES.BYTE,
+    },
+  ],
+};
+
+const VisualMaterialCharsField = ({ name }) => {
+  return (
+    <BytesField
+      name={name}
+      config={config}
+    />
+  );
+};
+
+VisualMaterialCharsField.propTypes = {
+  name: PropTypes.string.isRequired,
+};
+
+VisualMaterialCharsField.displayName = 'VisualMaterialCharsField';
+
+export default VisualMaterialCharsField;