+++ src/QuickMarcEditor/QuickMarcEditorRows/MaterialCharsField/index.js
@@ -1,0 +1,10 @@
+export * from './MaterialCharsFieldFactory';
+
+export { default as BookMaterialCharsField } from './BookMaterialCharsField';
+export { default as ComputerFileMaterialCharsField } from './ComputerFileMaterialCharsField';
+export { default as ContinuingMaterialCharsField } from './ContinuingMaterialCharsField';
+export { default as MapMaterialCharsField } from './MapMaterialCharsField';
+export { default as MixedMaterialCharsField } from './MixedMaterialCharsField';
+export { default as ScoreMaterialCharsField } from './ScoreMaterialCharsField';
+export { default as SoundMaterialCharsField } from './SoundMaterialCharsField';
+export { default as VisualMaterialCharsField } from './VisualMaterialCharsField';